#52 renamed product package into production
