+++ bakery/bakery-dao/src/test/java/sssvn/example/OrderTest.java
@@ -20,8 +20,8 @@ import sssvn.personnel.Carrier;
 import sssvn.personnel.CarrierCo;
 import sssvn.personnel.Person;
 import sssvn.personnel.validators.PersonInitialsValidator;
-import sssvn.product.Product;
-import sssvn.product.ProductCo;
+import sssvn.production.Product;
+import sssvn.production.ProductCo;
 import sssvn.test_config.AbstractDaoTestCase;
 import sssvn.test_config.UniversalConstantsForTesting;
 import ua.com.fielden.platform.entity.meta.MetaProperty;
