+++ bakery/bakery-pojo-bl/src/main/java/sssvn/config/ApplicationDomain.java
@@ -7,6 +7,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import sssvn.personnel.Person;
+import sssvn.production.Product;
 import ua.com.fielden.platform.basic.config.IApplicationDomainProvider;
 import ua.com.fielden.platform.domain.PlatformDomainTypes;
 import ua.com.fielden.platform.entity.AbstractEntity;
@@ -14,7 +15,6 @@ import sssvn.logistics.Location;
 import sssvn.logistics.Order;
 import sssvn.logistics.OrderItem;
 import sssvn.personnel.Manager;
-import sssvn.product.Product;
 import sssvn.personnel.Carrier;
 import sssvn.personnel.Employment;
