+++ bakery/bakery-pojo-bl/src/main/java/sssvn/logistics/validators/ProductAmountForOrderValidator.java
@@ -23,7 +23,7 @@ import ua.com.fielden.platform.test.ioc.UniversalConstantsForTesting;
 import ua.com.fielden.platform.types.Money;
 import ua.com.fielden.platform.utils.IUniversalConstants;
 import sssvn.logistics.OrderItem;
-import sssvn.product.Product;
+import sssvn.production.Product;
 import ua.com.fielden.platform.entity.meta.MetaProperty;
 import ua.com.fielden.platform.entity.meta.impl.AbstractBeforeChangeEventHandler;
 import ua.com.fielden.platform.error.Result;
