+++ bakery/bakery-web-server/src/main/java/sssvn/webapp/WebUiConfig.java
@@ -11,7 +11,7 @@ import sssvn.personnel.Carrier;
 import sssvn.personnel.Employment;
 import sssvn.personnel.Manager;
 import sssvn.personnel.Person;
-import sssvn.product.Product;
+import sssvn.production.Product;
 import sssvn.config.Modules;
 import sssvn.config.personnel.PersonWebUiConfig;
 import sssvn.webapp.config.logistics.LocationWebUiConfig;
@@ -20,7 +20,7 @@ import sssvn.webapp.config.logistics.OrderWebUiConfig;
 import sssvn.webapp.config.personnel.CarrierWebUiConfig;
 import sssvn.webapp.config.personnel.EmploymentWebUiConfig;
 import sssvn.webapp.config.personnel.ManagerWebUiConfig;
-import sssvn.webapp.config.product.ProductWebUiConfig;
+import sssvn.webapp.config.production.ProductWebUiConfig;
 import ua.com.fielden.platform.basic.config.Workflows;
 import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.reflection.TitlesDescsGetter;
